+++ backend/app.js
@@ -3,16 +3,25 @@ const app               = express();
 const path              = require('path');
 const mysql             = require('mysql');
 const Router            = require('./Router');
-const cors = require('cors');
+const cors              = require('cors');
 
 require('dotenv').config();
 
 app.use(express.static(path.join(__dirname, 'build')));
 app.use(express.json());
 
-app.use(cors({
-    origin: 'https://tedxoxford.co.uk'
-}));
+var whitelist = ['https://www.tedxoxford.com', 'http://localhost:3000']
+var corsOptions = {
+  origin: function (origin, callback) {
+    if (whitelist.indexOf(origin) !== -1) {
+      callback(null, true)
+    } else {
+      callback(new Error('Not allowed by CORS'))
+    }
+  }
+}
+
+app.use(cors(corsOptions));
 
 // Databases
 const db = mysql.createConnection({